uxen: change page_info count_info field to uint32_t
on 64b, this reduces the size of struct page_info from 24 byte
to 16 byte since it removes a 64 bit alignment requirement of
an unsigned long field

also disable DEBUG_MAPCACHE

+++ osx/uxen.kext/uxen_mem.c
@@ -982,7 +982,7 @@ release_page(uint32_t pfn)
 struct page_list_entry {
     uint32_t next, prev;
 #ifdef DEBUG
-    uintptr_t count_info;
+    uint32_t count_info;
 #endif
 };
